refactor: add repository interface
