+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/common/redis/message/service/ChatMessageService.java
@@ -2,7 +2,7 @@ package kr.co.pennyway.domain.common.redis.message.service;
 
 import kr.co.pennyway.common.annotation.DomainService;
 import kr.co.pennyway.domain.common.redis.message.domain.ChatMessage;
-import kr.co.pennyway.domain.common.redis.message.repository.ChatMessageRepository;
+import kr.co.pennyway.domain.common.redis.message.repository.ChatMessageRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -10,9 +10,9 @@ import lombok.extern.slf4j.Slf4j;
 @DomainService
 @RequiredArgsConstructor
 public class ChatMessageService {
-    private final ChatMessageRepository chatMessageRepository;
+    private final ChatMessageRepositoryImpl chatMessageRepositoryImpl;
 
     public ChatMessage save(ChatMessage chatMessage) {
-        return chatMessageRepository.save(chatMessage);
+        return chatMessageRepositoryImpl.save(chatMessage);
     }
 }
